Implement assignment of subject identifiers to existing users
This is for the use case where admins manually create OidcUsers. At this time, only the username is known. Now, if a user that has been manually created logs in for the first time, his subject identifier is assigned to the OidcUser entity. From this point on, the subject identifier cannot change. If it does, authentication will fail.