Make `PersistenceManagerFactory` more test-friendly
Currently, `createPersistenceManager` would recreate `pmf` on every invocation when the unit test mode is enabled.

Recreating `pmf` means validating the schema and recreating it if necessary, and takes a noticeable amount of time. This becomes a problem when a functionality is tested that calls `createPersistenceManager` multiple times, for example by using multiple `QueryManager` instances.

Instead a `pmf` instance should stay the same throughout a test case. Aftwards, it should be discarded, so it can be recreated in the next test case.

Additionally, when testing with external databases (e.g. via testcontainers), it must be possible to customize the `pmf` without the need of modifying environment variables. I added a public setter for `pmf`, which will fail when `pmf` was already initialized. It won't be possible to use it when deployed to a servlet context, where initialization happens via `contextInitialized`.

Signed-off-by: nscuro <nscuro@protonmail.com>