Add explicit plugin declaration for `maven-war-plugin`
It was only included in the `embedded-jetty` profile before. When building a Maven project of type `war`, the war plugin is executed per default in the `package` phase. In Maven 3.8.1 (bundled with latest IntelliJ), `maven-war-plugin` has version `2.2` per default, which is incompatible with Java 17.

This caused build failures when building without `embedded-jetty` profile, e.g. when running `mvn clean verify`. Adding `maven-war-plugin` explicitly with the desired version solves this issue.

Signed-off-by: nscuro <nscuro@protonmail.com>