Add support for protecting metrics with basic authentication
The metrics endpoint is intended to be used by machines, and access has to be available as soon as the application starts. While it's not viable to integrate this endpoint with Alpine's default AuthN/AuthZ model, basic auth is a good compromise.

Also, Prometheus supports basic auth, but not custom headers like `X-Api-Key` or similar.

Signed-off-by: nscuro <nscuro@protonmail.com>