Disable `javax.jdo.option.Multithreaded`
PMs are not designed to be used by multiple threads, and setting `javax.jdo.option.Multithreaded` is no guarantee for thread safety.

The expectation is that applications using Alpine are aware of this. `EntityManager`s in JPA / Hibernate are not thread safe either, so this concept is nothing new.

The locking logic employed by DN when the PM is multithreaded may cause deadlocks, even if it isn't used by multiple threads, see https://github.com/datanucleus/datanucleus-core/issues/479

Signed-off-by: nscuro <nscuro@protonmail.com>