Save and load secret key using encoded bytes format instead of Object Serialization
Using OS makes it impractical to provide a key that was generated before Alpine / the application using Alpine was first launched.

Secret keys can be loaded both in the old and the new format. This is done for backwards compatibility. Existing deployments of Alpine-based apps will not have to do anything and continue to function as before.

New keys will only be generated in the new encoded format. There is no automatic conversion of existing keys happening, because some users of Alpine may already have the key file set to read-only.

Additionally, the location of the secret key is now configurable. This is important to have when using k8s secrets, as nested mounts (e.g. `-v "./data:/data" -v "./secret.key:/data/.alpine/keys/secret.key"`) are an anti-pattern.

Closes #436