Allow customization of DataNucleus via pass-through properties
This change is mostly meant to address #493, but may be useful for other use cases as well.

Due to the sheer amount of configuration options in DataNucleus (and potentially other frameworks), it is not practical to add `AlpineKey`s for all of them.

With this change, it is possible to disable the DataNucleus L2 cache by either:

* Setting the property `alpine.datanucleus.cache.level2.type=none` in `application.properties`, or
* Setting the environment variable `ALPINE_DATANUCLEUS_CACHE_LEVEL2_TYPE=none`

Signed-off-by: nscuro <nscuro@protonmail.com>