oidc: default team
The assignment of (a) default team(s) happens once during user provisioning,
to give administrators a chance to change the teams for a user later.

The assignment of (a) default team(s) does not go through the indirection of
an OIDC Group, and expects the administrator to directly specify the teams
the user should be a member of. This is intentional, because otherwise it
could be confusing what would happen when the mapping between OIDC Groups
and Teams changes after a user was already provisioned. This also matches
the UI, which primarily shows (and allows changes to) the 'Team membership'
of a user on the 'OpenID Connect Users' page, rather than the OIDC Group
membership.

Fixes https://github.com/stevespringett/Alpine/issues/411