Track API key creation and last usage
In order to avoid introducing additional latency to requests, API key usage is buffered in-memory and asynchronously flushed to the database in intervals of 30s.

An API key counts as "used" if authentication with it succeeds.

Also, add support for an optional comment for API keys, as requested in #430 and #536.

Closes #362
Closes #430
Closes #536

Signed-off-by: nscuro <nscuro@protonmail.com>