Use lambdas for transaction scoping; Don't reload objects after commit
The way transactions were previously handled made it impossible to compose multiple transactional operations into larger tasks. `PersistenceManager#currentTransaction()#begin()` fails when the transaction is already active. This in turn opens the door to data inconsistencies, when parts of a task are committed but later operations fail, with no way to roll back to the original state before the task was started.

It also encouraged commits in high frequencies, which can take a toll on the database's write-ahead-log for high volume processing use-cases.

There was no handling of rollbacks, such that a failed commit could have "poisoned" an entire `PersistenceManager`, as no further persistence operations can be performed until after a rollback was executed.

This change introduces the `Transaction` class, which coordinates transaction begin, commit, and rollback. It is also capable of handling nested transactions, by joining an existing active one if required. This aspect can be controlled via propagation settings.

It also removes the need to reload objects from the database after each commit. This behavior has historically added quite a big overhead, and apparently was needed because DataNucleus transitions objects into "hollow" state (all fields except the primary key are unloaded). The unloading of fields is now disabled globally by setting the `DataNucleus.RetainValues` property to `true`.

The `ScopedCustomization` class was added to allow for scoped changes to a `PersistenceManager`, without impacting later operations on it. Usually things like adding fetch groups via `pm.getFetchPlan().addGroup("foo")` are intended to only affect one or a handful of queries, but in reality it affects all queries for the lifetime of the `PersistenceManager`. This has historically led to over-fetching. Now, it's easier to scope such things more tightly.

Further, this PR makes an effort to ensure that all `Query` objects are closed once consumed. Not closing query objects unnecessary leaks resources in cases where the owning `PersistenceManager` is **not** short-lived.

References:

* https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#lifecycle
* https://www.datanucleus.org/products/accessplatform_6_0/jdo/query.html#_closing_a_query
* https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#transaction_local
* https://github.com/search?q=repo%3AGoogleCloudPlatform%2Fdatanucleus-appengine%20retainvalues&type=code

Signed-off-by: nscuro <nscuro@protonmail.com>