Fix `getCount` for queries with subqueries
The `getCount` methods were refactored to not modify the original `Query` object in #552. Neither JDO nor DataNucleus offer a mechanism to copy a `Query` object 1:1, hence a new object was created instead.

This however broke `getCount` for queries that make use of subqueries (and potentially other advanced features).

This change reverts it back to temporarily modify the query object again.

Signed-off-by: nscuro <nscuro@protonmail.com>