Fix `NoClassDefFoundError: jakarta/servlet/ServletContext` in executable WAR
The `jakarta.servlet-api` JAR was not getting shaded into `alpine-executable-war` because it's scope defaulted to `provided`.

Removed the `scope` declaration in the parent POM. We define the `scope` explicitly in the respective Alpine modules now if it differs from `compile`.

Without the servlet API being present, the embedded Jetty failed to start with a `NoClassDefFoundError`.

Signed-off-by: nscuro <nscuro@protonmail.com>