Fix breaking change due to Jetty / Servlet API URI compliance enforcement
Due to a change in Jetty 12 and the Servlet API 6, it is no longer valid to pass URL-encoded values as path parameters. This is problematic for REST APIs such as Dependency-Track's, where this behavior causes endpoints to be unusable with certain parameters: https://github.com/DependencyTrack/dependency-track/issues/4238

Restore the legacy behavior of Jetty 9 for now. For the next major version bump of Alpine, this can be reversed to follow Servlet API 6 more strictly.

Signed-off-by: nscuro <nscuro@protonmail.com>