Enrich MDC with request method and request URI
Logs emitted during request handling can be ambiguous if a resource class contains many methods.

Access control failure logs are harder to interpret without the additional context of which endpoint was attempted to be accessed.

This change enriches the logger's MDC with both the request method, and the request URI.

If an endpoint accepts path parameters, the `requestUri` MDC variable will only contain the respective placeholder, *not* the parameter values themselves. This keeps values constant, making it easier to filter through logs.

Signed-off-by: nscuro <nscuro@protonmail.com>