Stop Jetty when context startup fails
Currently, the Jetty process is not stopped when the web app fails to start (e.g., due to an exception in a `ServletContextListener`).

While this behavior makes sense in settings where multiple apps are deployed in the same app server, it is unintuitive for embedded use cases since it gives the impression that the application is still running.

Signed-off-by: nscuro <nscuro@protonmail.com>